LIB-209 modified the data type of bookId from Long to Integer

+++ src/main/java/com/libraryportal/restapi/dao/ReviewRepository.java
@@ -11,6 +11,6 @@ public interface ReviewRepository extends JpaRepository<Review, Integer> {
 
     Page<Review> findBookReviewListByBookId(@RequestParam("bookId") Integer bookId, Pageable pageable);
     
-    Review findBookReviewByUserEmailAndBookId(String userEmail, Long bookId);
+    Review findBookReviewByUserEmailAndBookId(String userEmail, Integer bookId);
 
 }
